requirements.txt
elevenlabs
requests
python-dotenv
python-pptx
azure-cognitiveservices-speech
pyyaml
